Fixing type of type field from illumos-joyent bhyve-smbios tbl patch to meet smbiostbl spec (0-127)

+++ usr.sbin/bhyve/smbiostbl.c
@@ -836,7 +836,7 @@ smbios_parse(const char *opts)
 	char *lasts;
 	char *token;
 	char *end;
-	long type;
+	int type;
 	struct {
 		const char *key;
 		const char **targetp;
